requirements.txt
Flask
transformers
Pillow
pytest
